+++ jscomp/ext/ext_path.mli
@@ -1,5 +1,5 @@
 (* Copyright (C) 2017 Authors of ReScript
- * 
+ *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
@@ -17,7 +17,7 @@
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU Lesser General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA. *)
@@ -25,11 +25,11 @@
 type t
 
 val simple_convert_node_path_to_os_path : string -> string
-(** Js_output is node style, which means 
+(** Js_output is node style, which means
     separator is only '/'
 
-    if the path contains 'node_modules', 
-    [node_relative_path] will discard its prefix and 
+    if the path contains 'node_modules',
+    [node_relative_path] will discard its prefix and
     just treat it as a library instead
 *)
 
@@ -50,14 +50,14 @@ val combine : string -> string -> string
 val node_rebase_file : from:string -> to_:string -> string -> string
 
 val rel_normalized_absolute_path : from:string -> string -> string
-(** 
+(**
    TODO: could be highly optimized
-   if [from] and [to] resolve to the same path, a zero-length string is returned 
+   if [from] and [to] resolve to the same path, a zero-length string is returned
    Given that two paths are directory
 
-   A typical use case is 
+   A typical use case is
    {[
-     Filename.concat 
+     Filename.concat
        (rel_normalized_absolute_path cwd (Filename.dirname a))
        (Filename.basename a)
    ]}
@@ -69,7 +69,7 @@ val absolute_cwd_path : string -> string
 
 val concat : string -> string -> string
 (** [concat dirname filename]
-    The same as {!Filename.concat} except a tiny optimization 
+    The same as {!Filename.concat} except a tiny optimization
     for current directory simplification
 *)