+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Untagged variants: Support `bool`. https://github.com/rescript-lang/rescript-compiler/pull/6368
 - Support aliased types as payloads to untagged variants. https://github.com/rescript-lang/rescript-compiler/pull/6394
 - Support the async component for React Server Component in JSX V4. https://github.com/rescript-lang/rescript-compiler/pull/6399
+- Support `rescript.json` configuration file and deprecated `bsconfig.json`. https://github.com/rescript-lang/rescript-compiler/pull/6382
 
 #### :boom: Breaking Change
 - Add smart printer for pipe-chains.  https://github.com/rescript-lang/rescript-compiler/pull/6411 (the formatter will reformat existing code in certain cases)